pyproject.toml
[project]
name = "fourth-marketing-brain"
version = "0.1.0"
description = "Fourth's AI-powered marketing knowledge MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0.0,<3",
    "msal>=1.28.0",
    "httpx>=0.27.0",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.0",
    "python-pptx>=0.6.23",
    "azure-cosmos>=4.7.0",
    "azure-search-documents>=11.5.0",
    "aiohttp>=3.13.3",
]

[tool.hatch.build.targets.wheel]
packages = ["backends"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
